#12 refactor backend application and tests
* move models to separate folder
* split GameService into set of separate services
* add DI to tests
