+++ InfiniTicTacToe.Server/Controllers/GameController.cs
@@ -1,3 +1,4 @@
+using InfiniTicTacToe.Server.Models;
 using InfiniTicTacToe.Server.Services;
 using Microsoft.AspNetCore.Mvc;
 
@@ -5,12 +6,12 @@ namespace InfiniTicTacToe.Server.Controllers;
 
 [ApiController]
 [Route("api/game")]
-public class GameController(GameService gameService, ILogger<GameController> logger) : ControllerBase
+public class GameController(GameStorage gameStorage, ILogger<GameController> logger) : ControllerBase
 {
     [HttpGet("status")]
     public ActionResult<GameStatistics> GetCurrentStatus()
     {
-        var status = gameService.GetStats();
+        var status = gameStorage.GetStats();
         logger.LogInformation("Current status: {Games} games in progress", status.Games);
         return status;
     }
