+++ InfiniTicTacToe.Server/Models/GameScore.cs
@@ -1,0 +1,6 @@
+namespace InfiniTicTacToe.Server.Models;
+
+public sealed record GameScore(
+    string GameId,
+    int X,
+    int O);
