+++ InfiniTicTacToe.Server/Models/GameStatistic.cs
@@ -1,0 +1,6 @@
+namespace InfiniTicTacToe.Server.Models;
+
+public sealed record GameStatistics(
+    int Games,
+    int Players,
+    IReadOnlyCollection<GameScore> GameScores);
