+++ InfiniTicTacToe.Server/Models/GameStatus.cs
@@ -1,0 +1,8 @@
+namespace InfiniTicTacToe.Server.Models;
+
+public enum GameStatus
+{
+    Pending,
+    InProgress,
+    Finished,
+}
