+++ InfiniTicTacToe.Server/Models/Side.cs
@@ -1,0 +1,7 @@
+namespace InfiniTicTacToe.Server.Models;
+
+public enum Side
+{
+    X,
+    O,
+}
